Remove CA_Barstow.Daggett.AP.723815 from dependant housing characteristics.
After the respecification of Location based on county. CA_Barstow.Daggett.AP.723815 no longer has any units assigned to this weather station. This is due to the population centroid of San Bernardino County being close to the Los Angeles metro area, so the Long Beach location is assigned to San Bernadino County.